+++ projects/01_fyyur/starter_code/app.py
@@ -30,64 +30,10 @@ moment = Moment(app)
 app.config.from_object('config')
 db.init_app(app)
 db.app = app
-# db = SQLAlchemy(app)
-# from venue import db, venue_api
+
 app.register_blueprint(venue_api, url_prefix='/venues')
 migrate = Migrate(app, db, compare_type=True)
 
-# #----------------------------------------------------------------------------#
-# # Models.
-# #----------------------------------------------------------------------------#
-
-# class Venue(db.Model):
-#     __tablename__ = 'Venue'
-
-#     id = db.Column(db.Integer, primary_key=True)
-#     name = db.Column(db.String)
-#     city = db.Column(db.String(120))
-#     state = db.Column(db.String(120))
-#     address = db.Column(db.String(120))
-#     phone = db.Column(db.String(120))
-#     website = db.Column(db.String())
-#     image_link = db.Column(db.String(500))
-#     facebook_link = db.Column(db.String(120))
-#     genres = db.Column(db.ARRAY(db.String(120)))
-#     seeking_talent = db.Column(db.Boolean, default=False)
-#     seeking_description = db.Column(db.String, nullable=True)
-#     shows = db.relationship('Show', backref='venue', lazy='dynamic')
-
-# class Artist(db.Model):
-#     __tablename__ = 'Artist'
-
-#     id = db.Column(db.Integer, primary_key=True)
-#     name = db.Column(db.String)
-#     city = db.Column(db.String(120))
-#     state = db.Column(db.String(120))
-#     phone = db.Column(db.String(120))
-#     website = db.Column(db.String())
-#     image_link = db.Column(db.String(500))
-#     facebook_link = db.Column(db.String(120))
-#     genres = db.Column(db.ARRAY(db.String(120)))
-#     seeking_venue = db.Column(db.Boolean, default=False)
-#     seeking_description = db.Column(db.String, nullable=True)
-#     shows = db.relationship('Show', backref='artist', lazy='dynamic')
-
-# class Show(db.Model):
-#   __tablename__ = 'Show'
-#   id = db.Column(db.Integer, primary_key=True)
-#   start_time = db.Column(db.DateTime, nullable=False)
-#   artist_id = db.Column(db.Integer, db.ForeignKey('Artist.id'), nullable=False)
-#   venue_id = db.Column(db.Integer, db.ForeignKey('Venue.id'), nullable=False)
-  
-#   @hybrid_property
-#   def isUpcoming(self):
-#     now = datetime.now(self.start_time.tzinfo)
-#     return (self.start_time > now)
-
-#   @isUpcoming.expression
-#   def isUpcoming(cls):
-#     now = datetime.now()
-#     return (cls.start_time > now)
 
 #----------------------------------------------------------------------------#
 # Filters.
@@ -118,99 +64,6 @@ app.jinja_env.filters['datetime'] = format_datetime
 @app.route('/')
 def index():
   return render_template('pages/home.html')
-
-
-# #  Venues
-# #  ----------------------------------------------------------------
-
-def groupVenuesByCityAndState():
-  return groupClassBy(Venue, 'venues', 'city', 'state')
-
-# @app.route('/venues')
-# def venues():
-#   return "venues()"
-# def venues():
-#   data = groupVenuesByCityAndState()
-#   return render_template('pages/venues.html', areas=data);
-
-@app.route('/venues/search', methods=['POST'])
-def search_venues():
-  # TODO: implement search on artists with partial string search. Ensure it is case-insensitive.
-  # seach for Hop should return "The Musical Hop".
-  # search for "Music" should return "The Musical Hop" and "Park Square Live Music & Coffee"
-  response={
-    "count": 1,
-    "data": [{
-      "id": 2,
-      "name": "The Dueling Pianos Bar",
-      "num_upcoming_shows": 0,
-    }]
-  }
-  return render_template('pages/search_venues.html', results=response, search_term=request.form.get('search_term', ''))
-
-# @app.route('/venues/<int:venue_id>')
-# def show_venue(venue_id):
-#   venue = Venue.query.get(venue_id)
-#   upcoming_shows= venue.shows.filter(Show.isUpcoming).all()
-#   past_shows= venue.shows.filter(Show.isUpcoming == False).all()
-#   return render_template('pages/show_venue.html', venue=venue,
-#     upcoming_shows=upcoming_shows,
-#     past_shows=past_shows)
-
-#  Create Venue
-#  ----------------------------------------------------------------
-
-@app.route('/venues/create', methods=['GET'])
-def create_venue_form():
-  form = VenueForm()
-  return render_template('forms/new_venue.html', form=form)
-
-def populateClassObject(className, request):
-  obj = className()
-  attributes = inspect.getmembers(className, lambda a:not(inspect.isroutine(a)))
-  attributes = [a[0] for a in attributes if not(a[0].startswith('_'))]
-  for a in attributes:
-    if a in request.form:
-      if isinstance(getattr(className, a).type, db.ARRAY):
-        setattr(obj, a, request.form.getlist(a))
-      else:
-        setattr(obj, a, request.form[a])
-  return obj
-
-def populateObjectFromRequest(obj, request):
-  attributes = inspect.getmembers(obj, lambda a:not(inspect.isroutine(a)))
-  attributes = [a[0] for a in attributes if not(a[0].startswith('_'))]
-  for a in attributes:
-    if a in request.form:
-      if isinstance(getattr(type(obj), a).type, db.ARRAY):
-        setattr(obj, a, request.form.getlist(a))
-      else:
-        setattr(obj, a, request.form[a])
-  return obj
-
-@app.route('/venues/create', methods=['POST'])
-def create_venue_submission():
-  try:
-    venue = Venue()
-    venue = populateObjectFromRequest(venue, request)
-    db.session.add(venue)
-    db.session.commit()
-    flash('Venue ' + request.form['name'] + ' was successfully listed!')
-  except:
-    db.session.rollback()
-    flash('An error occurred. Venue ' + request.form['name'] + ' could not be listed.')
-  finally:
-    db.session.close()
-  return render_template('pages/home.html')
-
-@app.route('/venues/<venue_id>', methods=['DELETE'])
-def delete_venue(venue_id):
-  # TODO: Complete this endpoint for taking a venue_id, and using
-  # SQLAlchemy ORM to delete a record. Handle cases where the session commit could fail.
-
-  # BONUS CHALLENGE: Implement a button to delete a Venue on a Venue Page, have it so that
-  # clicking that button delete it from the db then redirect the user to the homepage
-  return None
 
 #  Artists
 #  ----------------------------------------------------------------
@@ -245,13 +98,7 @@ def show_artist(artist_id):
 
 #  Update
 #  ----------------------------------------------------------------
-def setFormDefaultValues(form, obj):
-  attributes = inspect.getmembers(form, lambda a:not(inspect.isroutine(a)))
-  attributes = [a[0] for a in attributes if not(a[0].startswith('_'))]
-  for a in attributes:
-    if hasattr(obj, a):
-      setattr(getattr(form, a), "default", getattr(obj, a))
-  form.process()
+
 
 @app.route('/artists/<int:artist_id>/edit', methods=['GET'])
 def edit_artist(artist_id):
@@ -274,26 +121,6 @@ def edit_artist_submission(artist_id):
     db.session.close()
   return redirect(url_for('show_artist', artist_id=artist_id))
 
-@app.route('/venues/<int:venue_id>/edit', methods=['GET'])
-def edit_venue(venue_id):
-  form = VenueForm()
-  venue = Venue.query.get(venue_id) 
-  setFormDefaultValues(form, venue)
-  return render_template('forms/edit_venue.html', form=form, venue=venue)
-
-@app.route('/venues/<int:venue_id>/edit', methods=['POST'])
-def edit_venue_submission(venue_id):
-  try:
-    venue = Venue.query.get(venue_id)
-    venue = populateObjectFromRequest(venue, request)
-    db.session.commit()
-    flash('Venue ' + request.form['name'] + ' was successfully edited!')
-  except:
-    db.session.rollback()
-    flash('An error occurred. Venue ' + request.form['name'] + ' could not be edited.')
-  finally:
-    db.session.close()
-  return redirect(url_for('show_venue', venue_id=venue_id))
 
 #  Create Artist
 #  ----------------------------------------------------------------
